Refactor finishOrNext method of DynamicTabAdapterStrategy
