+++ app/src/cambodia/java/org/eyeseetea/malariacare/layout/adapters/survey/strategies/DynamicTabAdapterStrategy.java
@@ -1,30 +1,18 @@
 package org.eyeseetea.malariacare.layout.adapters.survey.strategies;
 
 
-import android.os.Handler;
-
-import org.eyeseetea.malariacare.BuildConfig;
-import org.eyeseetea.malariacare.DashboardActivity;
 import org.eyeseetea.malariacare.data.database.model.Question;
 import org.eyeseetea.malariacare.data.database.model.Survey;
-import org.eyeseetea.malariacare.data.database.model.Value;
-import org.eyeseetea.malariacare.data.database.utils.PreferencesState;
-import org.eyeseetea.malariacare.data.database.utils.Session;
-import org.eyeseetea.malariacare.domain.entity.Validation;
 import org.eyeseetea.malariacare.layout.adapters.survey.DynamicTabAdapter;
-import org.eyeseetea.malariacare.strategies.ReviewFragmentStrategy;
-import org.eyeseetea.malariacare.strategies.UIMessagesStrategy;
 import org.eyeseetea.malariacare.utils.Constants;
 import org.eyeseetea.malariacare.views.question.IQuestionView;
 
 import java.util.List;
 
-public class DynamicTabAdapterStrategy implements IDynamicTabAdapterStrategy {
-
-    DynamicTabAdapter mDynamicTabAdapter;
+public class DynamicTabAdapterStrategy extends ADynamicTabAdapterStrategy {
 
     public DynamicTabAdapterStrategy(DynamicTabAdapter dynamicTabAdapter) {
-        this.mDynamicTabAdapter = dynamicTabAdapter;
+        super(dynamicTabAdapter);
     }
 
     @Override
@@ -70,45 +58,4 @@ public class DynamicTabAdapterStrategy implements IDynamicTabAdapterStrategy {
         return;
     }
 
-    @Override
-    public void finishOrNext() {
-        try {
-            System.out.println(Session.getMalariaSurvey().getValuesFromDB().toString());
-            System.out.println(Session.getStockSurvey().getValuesFromDB().toString());
-        } catch (Exception e) {
-        }
-        if (Validation.hasErrors()) {
-            Validation.showErrors();
-            DynamicTabAdapter.setIsClicked(false);
-            return;
-        }
-        if (mDynamicTabAdapter.navigationController.getCurrentQuestion().hasCompulsoryNotAnswered
-                ()) {
-
-            UIMessagesStrategy.getInstance().showCompulsoryUnansweredToast();
-            DynamicTabAdapter.setIsClicked(false);
-            return;
-        }
-        final Handler handler = new Handler();
-        handler.postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                Question question = mDynamicTabAdapter.navigationController.getCurrentQuestion();
-                Value value = question.getValueBySession();
-                if (mDynamicTabAdapter.isDone(value)) {
-                    mDynamicTabAdapter.navigationController.isMovingToForward = false;
-                    if (!ReviewFragmentStrategy.shouldShowReviewScreen() || !BuildConfig.reviewScreen) {
-                        mDynamicTabAdapter.surveyShowDone();
-                    } else {
-                        DashboardActivity.dashboardActivity.showReviewFragment();
-                        mDynamicTabAdapter.hideKeyboard(
-                                PreferencesState.getInstance().getContext());
-                        DynamicTabAdapter.setIsClicked(false);
-                    }
-                    return;
-                }
-                mDynamicTabAdapter.next();
-            }
-        }, 750);
-    }
 }
