+++ app/src/myanmar/java/org/eyeseetea/malariacare/layout/adapters/survey/strategies/DynamicTabAdapterStrategy.java
@@ -3,10 +3,6 @@ package org.eyeseetea.malariacare.layout.adapters.survey.strategies;
 
 import static org.eyeseetea.malariacare.data.database.utils.Session.getMalariaSurvey;
 
-import android.os.Handler;
-
-import org.eyeseetea.malariacare.BuildConfig;
-import org.eyeseetea.malariacare.DashboardActivity;
 import org.eyeseetea.malariacare.R;
 import org.eyeseetea.malariacare.data.database.model.Option;
 import org.eyeseetea.malariacare.data.database.model.Question;
@@ -16,11 +12,8 @@ import org.eyeseetea.malariacare.data.database.utils.PreferencesState;
 import org.eyeseetea.malariacare.data.database.utils.Session;
 import org.eyeseetea.malariacare.domain.entity.Treatment;
 import org.eyeseetea.malariacare.domain.entity.TreatmentQueries;
-import org.eyeseetea.malariacare.domain.entity.Validation;
 import org.eyeseetea.malariacare.layout.adapters.survey.DynamicTabAdapter;
 import org.eyeseetea.malariacare.layout.listeners.question.QuestionAnswerChangedListener;
-import org.eyeseetea.malariacare.strategies.ReviewFragmentStrategy;
-import org.eyeseetea.malariacare.strategies.UIMessagesStrategy;
 import org.eyeseetea.malariacare.utils.Constants;
 import org.eyeseetea.malariacare.utils.GradleVariantConfig;
 import org.eyeseetea.malariacare.views.question.IQuestionView;
@@ -32,16 +25,14 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
-public class DynamicTabAdapterStrategy implements IDynamicTabAdapterStrategy {
-
-    DynamicTabAdapter mDynamicTabAdapter;
+public class DynamicTabAdapterStrategy extends ADynamicTabAdapterStrategy {
     /**
      * Added to save the dose by a quetion id when questions are dynamic treatment questions
      */
     HashMap<Long, Float> doseByQuestion;
 
     public DynamicTabAdapterStrategy(DynamicTabAdapter dynamicTabAdapter) {
-        this.mDynamicTabAdapter = dynamicTabAdapter;
+        super(dynamicTabAdapter);
     }
 
     @Override
@@ -172,47 +163,5 @@ public class DynamicTabAdapterStrategy implements IDynamicTabAdapterStrategy {
                     new QuestionAnswerChangedListener(dynamicTabAdapter,
                             !GradleVariantConfig.isButtonNavigationActive()));
         }
-    }
-
-    @Override
-    public void finishOrNext() {
-        try {
-            System.out.println(Session.getMalariaSurvey().getValuesFromDB().toString());
-            System.out.println(Session.getStockSurvey().getValuesFromDB().toString());
-        } catch (Exception e) {
-        }
-        if (Validation.hasErrors()) {
-            Validation.showErrors();
-            DynamicTabAdapter.setIsClicked(false);
-            return;
-        }
-        if (mDynamicTabAdapter.navigationController.getCurrentQuestion().hasCompulsoryNotAnswered
-                ()) {
-
-            UIMessagesStrategy.getInstance().showCompulsoryUnansweredToast();
-            DynamicTabAdapter.setIsClicked(false);
-            return;
-        }
-        final Handler handler = new Handler();
-        handler.postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                Question question = mDynamicTabAdapter.navigationController.getCurrentQuestion();
-                Value value = question.getValueBySession();
-                if (mDynamicTabAdapter.isDone(value)) {
-                    mDynamicTabAdapter.navigationController.isMovingToForward = false;
-                    if (!ReviewFragmentStrategy.shouldShowReviewScreen() || !BuildConfig.reviewScreen) {
-                        mDynamicTabAdapter.surveyShowDone();
-                    } else {
-                        DashboardActivity.dashboardActivity.showReviewFragment();
-                        mDynamicTabAdapter.hideKeyboard(
-                                PreferencesState.getInstance().getContext());
-                        DynamicTabAdapter.setIsClicked(false);
-                    }
-                    return;
-                }
-                mDynamicTabAdapter.next();
-            }
-        }, 750);
     }
 }
